requirements.txt
# Core Data Processing
pyspark==3.3.0
delta-spark==2.1.0
pandas==1.5.3
numpy==1.24.3
pyarrow==12.0.0

# AWS SDK
boto3==1.26.137
awswrangler==3.1.1
s3fs==2023.5.0

# Airflow
apache-airflow==2.6.1
apache-airflow-providers-amazon==8.1.0
apache-airflow-providers-postgres==5.5.1

# Data Quality
great-expectations==0.16.13
pydantic==1.10.8

# Database Connectors
sqlalchemy==1.4.48
psycopg2-binary==2.9.6
redshift-connector==2.0.913

# Testing
pytest==7.3.1
pytest-cov==4.1.0
pytest-mock==3.10.0
moto==4.1.10  # AWS mocking

# Code Quality
black==23.3.0
pylint==2.17.4
mypy==1.3.0
isort==5.12.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0
click==8.1.3
loguru==0.7.0

# Notebook Support
jupyter==1.0.0
notebook==6.5.4
ipykernel==6.23.1
matplotlib==3.7.1
seaborn==0.12.2

# Documentation
sphinx==6.2.1
sphinx-rtd-theme==1.2.1